+++ api-builder-plugin-fn-file/test/testStandardFileHandling.js
@@ -164,11 +164,18 @@ describe('Standard file handling test', () => {
 			expect(output).to.equal('error');
 		});
 
-		it('should fail with an unknown file', async () => {
-			const { value, output } = await flowNode.readFile({ filename: 'UnknownFile' });
+		it('should fail with an unknown file when it should be treated as an error', async () => {
+			const { value, output } = await flowNode.readFile({ filename: 'UnknownFile', notFoundFails: true });
 
 			expect(value.message).to.match(/.*Error reading file: UnknownFile.*/)
 			expect(output).to.equal('error');
+		});
+
+		it.only('should fail with an unknown file when it should be treated as an error', async () => {
+			const { value, output } = await flowNode.readFile({ filename: 'UnknownFile' });
+
+			expect(output).to.equal('notFound');
+			expect(value).to.match(/.*File: UnknownFile not found.*/)
 		});
 
 		it('should read a standard file as normal text', async () => {